feat(asb): Listen for encrypted signature in XmrLocked state
If we send Bob the transfer proof, but for whatever reason we do not receive an acknoledgement from him we would be stuck in this state forever (deadlock). By listening for the encrypted signature here we can still proceed to the next state even if Bob does not respond with an acknoledgement.