fix monero integration test harness
Integration tests were failing because the wallet/
node couldn't keep up with the amount of blocks we
generated and took some time before it even reali-
zed it was out of sync.
That lead to unrecognized balances.

Fix: we wait 0.25 seconds between each new block.
This gives the wallet/node enough time to realize
it's out of sync.